chore: move date helpers to utils package
