+++ packages/date-publication/package.json
@@ -21,7 +21,7 @@
       "node_modules/(?!@times-components)/"
     ],
     "testMatch": [
-      "<rootDir>/packages/date-publication/**.test.js"
+      "<rootDir>/packages/date-publication/__tests__/*.test.js"
     ]
   },
   "repository": {
@@ -63,6 +63,7 @@
     "webpack": "3.3.0"
   },
   "dependencies": {
+    "@times-components/utils": "0.0.11",
     "date-fns": "1.28.5",
     "prop-types": "15.6.0"
   },
