+++ packages/utils/__tests__/strings.test.js
@@ -1,6 +1,6 @@
 /* eslint-env jest */
 
-import { ratioTextToFloat } from "./strings";
+import { ratioTextToFloat } from "../strings";
 
 describe("ratioTextToFloat should", () => {
   it("return 1 if no string is given", () => {
